+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.3 - 2025-04-13
+
+### Added
+- Fix Keyboard Navigation
+- add rotate & rotateto method to controller
+- add centeronrect method to controller
+- update center method (viewportsize and contentsize are no longer required in it)
+- introduced a bunch of flags like 'enableZoom'
+
 ## 0.0.2 - 2024-04-12
 
 ### Added